requirements.txt
fastapi
uvicorn[standard]
web3
python-dotenv
shapely
pyproj
pydantic